feat: Update e2e tests to run on new canvas (no-changelog) (#12784)
