+++ .github/workflows/e2e-reusable.yml
@@ -41,11 +41,6 @@ on:
         description: 'PR number to run tests for.'
         required: false
         type: number
-      node_view_version:
-        description: 'Node View version to run tests with.'
-        required: false
-        default: '1'
-        type: string
     secrets:
       CYPRESS_RECORD_KEY:
         description: 'Cypress record key.'
@@ -165,7 +160,7 @@ jobs:
           spec: '${{ inputs.spec }}'
         env:
           NODE_OPTIONS: --dns-result-order=ipv4first
-          CYPRESS_NODE_VIEW_VERSION: ${{ inputs.node_view_version }}
+          CYPRESS_NODE_VIEW_VERSION: 2
           CYPRESS_RECORD_KEY: ${{ secrets.CYPRESS_RECORD_KEY }}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           E2E_TESTS: true
