+++ cypress/e2e/12-canvas.cy.ts
@@ -28,8 +28,6 @@ const ZOOM_IN_X2_FACTOR = 1.5625; // Zoom in factor after two clicks
 const ZOOM_OUT_X1_FACTOR = 0.8;
 const ZOOM_OUT_X2_FACTOR = 0.64;
 
-const PINCH_ZOOM_IN_FACTOR = 1.05702;
-const PINCH_ZOOM_OUT_FACTOR = 0.946058;
 const RENAME_NODE_NAME = 'Something else';
 const RENAME_NODE_NAME2 = 'Something different';
 
@@ -369,26 +367,6 @@ describe('Canvas Node Manipulation and Navigation', () => {
 			zoomAndCheck('zoomOut', ZOOM_OUT_X2_FACTOR);
 		});
 
-		it('should zoom using scroll or pinch gesture', () => {
-			WorkflowPage.actions.pinchToZoom(1, 'zoomIn');
-
-			// V2 Canvas is using the same zoom factor for both pinch and scroll
-			cy.ifCanvasVersion(
-				() => checkZoomLevel(PINCH_ZOOM_IN_FACTOR),
-				() => checkZoomLevel(ZOOM_IN_X1_FACTOR),
-			);
-
-			WorkflowPage.actions.pinchToZoom(1, 'zoomOut');
-			checkZoomLevel(1); // Zoom in 1x + Zoom out 1x should reset to default (=1)
-
-			WorkflowPage.actions.pinchToZoom(1, 'zoomOut');
-
-			cy.ifCanvasVersion(
-				() => checkZoomLevel(PINCH_ZOOM_OUT_FACTOR),
-				() => checkZoomLevel(ZOOM_OUT_X1_FACTOR),
-			);
-		});
-
 		it('should reset zoom', () => {
 			WorkflowPage.getters.resetZoomButton().should('not.exist');
 			WorkflowPage.getters.zoomInButton().click();
@@ -542,36 +520,5 @@ describe('Canvas Node Manipulation and Navigation', () => {
 			cy.get('[class*=hasIssues]').should('have.length', 1);
 			NDVDialog.actions.close();
 		});
-	});
-
-	// FIXME: Canvas V2: Unknown nodes should still render connection endpoints
-	it('should render connections correctly if unkown nodes are present', () => {
-		const unknownNodeName = 'Unknown node';
-		cy.createFixtureWorkflow('workflow-with-unknown-nodes.json', 'Unknown nodes');
-
-		WorkflowPage.getters.canvasNodeByName(`${unknownNodeName} 1`).should('exist');
-		WorkflowPage.getters.canvasNodeByName(`${unknownNodeName} 2`).should('exist');
-		WorkflowPage.actions.zoomToFit();
-
-		cy.draganddrop(
-			WorkflowPage.getters.getEndpointSelector('plus', `${unknownNodeName} 1`),
-			WorkflowPage.getters.getEndpointSelector('input', EDIT_FIELDS_SET_NODE_NAME),
-		);
-
-		cy.draganddrop(
-			WorkflowPage.getters.getEndpointSelector('plus', `${unknownNodeName} 2`),
-			WorkflowPage.getters.getEndpointSelector('input', `${EDIT_FIELDS_SET_NODE_NAME}1`),
-		);
-
-		WorkflowPage.actions.executeWorkflow();
-		cy.contains('Unrecognized node type').should('be.visible');
-
-		WorkflowPage.actions.deselectAll();
-		WorkflowPage.actions.deleteNodeFromContextMenu(`${unknownNodeName} 1`);
-		WorkflowPage.actions.deleteNodeFromContextMenu(`${unknownNodeName} 2`);
-
-		WorkflowPage.actions.executeWorkflow();
-
-		cy.contains('Unrecognized node type').should('not.exist');
 	});
 });
