--- cypress/e2e/1858-PAY-can-use-context-menu.ts
@@ -1,21 +1,0 @@
-import { WorkflowPage as WorkflowPageClass } from '../pages/workflow';
-
-const WorkflowPage = new WorkflowPageClass();
-
-describe('PAY-1858 context menu', () => {
-	it('can use context menu on saved workflow', () => {
-		WorkflowPage.actions.visit();
-		cy.createFixtureWorkflow('Test_workflow_filter.json', 'test');
-
-		WorkflowPage.getters.canvasNodes().should('have.length', 5);
-		WorkflowPage.actions.deleteNodeFromContextMenu('Then');
-		WorkflowPage.getters.canvasNodes().should('have.length', 4);
-
-		WorkflowPage.actions.hitSaveWorkflow();
-
-		cy.reload();
-		WorkflowPage.getters.canvasNodes().should('have.length', 4);
-		WorkflowPage.actions.deleteNodeFromContextMenu('Code');
-		WorkflowPage.getters.canvasNodes().should('have.length', 3);
-	});
-});
