+++ cypress/support/e2e.ts
@@ -38,7 +38,21 @@ beforeEach(() => {
 		data: { status: 'success', message: 'Tested successfully' },
 	}).as('credentialTest');
 
-	cy.intercept('POST', '/rest/license/renew', {});
+	cy.intercept('POST', '/rest/license/renew', {
+		data: {
+			usage: {
+				activeWorkflowTriggers: {
+					limit: -1,
+					value: 0,
+					warningThreshold: 0.8,
+				},
+			},
+			license: {
+				planId: '',
+				planName: 'Community',
+			},
+		},
+	});
 
 	cy.intercept({ pathname: '/api/health' }, { status: 'OK' }).as('healthCheck');
 	cy.intercept({ pathname: '/api/versions/*' }, [
