+++ packages/editor-ui/src/views/NodeView.v2.vue
@@ -349,7 +349,9 @@ async function initializeRoute(force = false) {
 		if (!isAlreadyInitialized) {
 			historyStore.reset();
 
-			await loadCredentials();
+			if (!isDemoRoute.value) {
+				await loadCredentials();
+			}
 
 			// If there is no workflow id, treat it as a new workflow
 			if (isNewWorkflowRoute.value || !workflowId.value) {
